feat(sdk-dotnet): Refactor classes to avoid having a class helper
