+++ sdk-dotnet/LittleHorse.Sdk.Tests/Workflow/Spec/LHVariableAssigmentTest.cs
@@ -1,17 +1,16 @@
 using System;
 using LittleHorse.Sdk.Common.Proto;
-using LittleHorse.Sdk.Helper;
 using LittleHorse.Sdk.Workflow.Spec;
 using Moq;
 using Xunit;
 
 namespace LittleHorse.Sdk.Tests.Helper;
 
-public class LHVariableAssigmentHelperTest
+public class LHVariableAssigmentTest
 {
     private WorkflowThread _parentWfThread;
     
-    public LHVariableAssigmentHelperTest()
+    public LHVariableAssigmentTest()
     {
         LHLoggerFactoryProvider.Initialize(null);
         var workflowName = "TestWorkflow";
@@ -23,7 +22,7 @@ public class LHVariableAssigmentHelperTest
     [Fact]
     public void VariableAssigment_WithNoVariable_ShouldReturnNoneLiteralValue()
     {
-        var variableAssigment = LHVariableAssigmentHelper.AssignVariable(null);
+        var variableAssigment = _parentWfThread.AssignVariableHelper(null);
         
         Assert.Equal(VariableValue.ValueOneofCase.None, variableAssigment.LiteralValue.ValueCase);
         Assert.Equal(String.Empty, variableAssigment.VariableName);
@@ -34,7 +33,7 @@ public class LHVariableAssigmentHelperTest
     {
         var wfRunVariable = new WfRunVariable("TestVariable", VariableType.Str, _parentWfThread);
         
-        var variableAssigment = LHVariableAssigmentHelper.AssignVariable(wfRunVariable);
+        var variableAssigment = _parentWfThread.AssignVariableHelper(wfRunVariable);
         
         Assert.Equal(wfRunVariable.Name, variableAssigment.VariableName);
         Assert.Equal(String.Empty, variableAssigment.JsonPath);
@@ -46,7 +45,7 @@ public class LHVariableAssigmentHelperTest
         var wfRunVariable = new WfRunVariable("TestVariable", VariableType.JsonObj, _parentWfThread);
         var wfRunVariableWithJson = wfRunVariable.WithJsonPath("$.order");
 
-        var variableAssigment = LHVariableAssigmentHelper.AssignVariable(wfRunVariableWithJson);
+        var variableAssigment = _parentWfThread.AssignVariableHelper(wfRunVariableWithJson);
         
         Assert.Equal(wfRunVariableWithJson.Name, variableAssigment.VariableName);
         Assert.Equal(wfRunVariableWithJson.JsonPath, variableAssigment.JsonPath);
@@ -58,7 +57,7 @@ public class LHVariableAssigmentHelperTest
        var nodeOutput = new NodeOutput("wait-to-collect-order-data", _parentWfThread);
        nodeOutput.JsonPath = "$.order";
        
-       var variableAssigment = LHVariableAssigmentHelper.AssignVariable(nodeOutput);
+       var variableAssigment = _parentWfThread.AssignVariableHelper(nodeOutput);
        
        Assert.Equal(nodeOutput.NodeName, variableAssigment.NodeOutput.NodeName);
        Assert.Equal(nodeOutput.JsonPath, variableAssigment.JsonPath);
@@ -71,7 +70,7 @@ public class LHVariableAssigmentHelperTest
     [InlineData(7.892)]
     public void VariableAssigment_WithNotDefinedObject_ShouldAssignObjectAsDefaultVariable(object notDefinedObject)
     {
-        var variableAssigment = LHVariableAssigmentHelper.AssignVariable(notDefinedObject);
+        var variableAssigment = _parentWfThread.AssignVariableHelper(notDefinedObject);
         
         Assert.Contains(notDefinedObject.ToString()!.ToLower(), variableAssigment.LiteralValue.ToString().ToLower());
     }
